+++ HISTORY.md
@@ -19,6 +19,7 @@ v0.5 - 2014/05/16
 * New ContentOwner subclass of Account with access to partnered channels
 * Automatically refresh the access token when it expires or becomes invalid
 * Retry once YouTube earning queries that return error 503
+* Wait 3 seconds and retry *every* request that returns 500, 503 or 400 with "Invalid query"
 
 v0.4 - 2014/05/09
 --------------------